+++ src/main/kotlin/com/github/oxc/project/oxcintellijplugin/Constants.kt
@@ -1,0 +1,3 @@
+package com.github.oxc.project.oxcintellijplugin
+
+const val NOTIFICATION_GROUP = "Oxc"